// src/app/lib/generated/models/Personality.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Personality = {
    Caregiver: 'caregiver',
    Shy: 'shy',
    Submissive: 'submissive',
    Mean: 'mean',
    Dominant: 'dominant',
    Flirty: 'flirty',
    Lover: 'lover',
    Nympho: 'nympho'
} as const;
export type Personality = typeof Personality[keyof typeof Personality];


export function instanceOfPersonality(value: any): boolean {
    for (const key in Personality) {
        if (Object.prototype.hasOwnProperty.call(Personality, key)) {
            if (Personality[key as keyof typeof Personality] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PersonalityFromJSON(json: any): Personality {
    return PersonalityFromJSONTyped(json, false);
}

export function PersonalityFromJSONTyped(json: any, ignoreDiscriminator: boolean): Personality {
    return json as Personality;
}

export function PersonalityToJSON(value?: Personality | null): any {
    return value as any;
}

export function PersonalityToJSONTyped(value: any, ignoreDiscriminator: boolean): Personality {
    return value as Personality;
}